Reverting to use LINARO 2015.03 GCC 4.9.3 with AUTO FPU.
GCC 5.1 maybe not best solution for 5.0.2 kernel.

I suspect that RAM bugs that brake kernel in low RAM conditions
started when i have changed GCC. before that all was OK.

But much changed in code, any way it's good to test this.

with LINARO all was great with KK Kernel.